module input: Refactor 'relevance' to 'implicit'
The term 'implicit' better represents what this parameter actually does.

If implicit is `True` a module is activated if any module in it's input or
output chain is activated during the given cue.

If implicit is `False` a module is only activated if it is explicitly activated
in a given cue.