cues: Don't activate explicit dependency module
EFFECT
======

Before this patch modules were activated if they were explicit inputs
of an explicitly activated module.

So in the following example cue 1 activated both the sine and the mixer
module:

'''
[configure.module.mixer.master]
send_to_physical_output                 = true
audio_input_0                           = "sine.0"

[configure.module.sine.0]

[cue.1.mixer.master]
'''

After this patch cue 1 only activates the mixer module, but doesn't
activate the sine module.

RATIONALE
=========

This change is necessary in order to allow modules to activate or
deactivate its inputs. This is needed to implement an event module,
which triggers the activation of its input modules (basically a meta
module).

But this change actually also makes sense in the whole walkman setup.
The implicit activation of explicit input module never really made
sense, it's better to stay explicit if it has already been declared so
(this also makes the difference between explicit and implicit inputs
much more clear:

    - if explicit: this module is only activated if it is explicitly
      activated

    - if implicit: this module is activated if any other activated
      module depends on it

)