mixins: Allow multiple pyo_object in PyoObjectMixin
Let's consider a patch bay where four different audio streams are
processed by the same audio effect with the same settings. But
after this audio effect they all get very different audio effects again.

In current walkman we only have the option to define the same module
four times with the same values, because otherwise it would always
get mixed.

So what we need to only specify the settings at the same is to allow
modules to have multiple pyo objects within themselves.
The first step to do this is to allow PyoObjectMixin to have multiple
pyo objects. This patch does exactly this.